refactor: spring-jdbc 적용
