+++ build.gradle
@@ -22,6 +22,8 @@ dependencies {
 
     implementation 'net.rakugakibox.spring.boot:logback-access-spring-boot-starter:2.7.1'
     implementation 'pl.allegro.tech.boot:handlebars-spring-boot-starter:0.3.4'
+    implementation 'junit:junit:4.13.1'
+    implementation 'junit:junit:4.13.1'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'io.rest-assured:rest-assured:4.4.0'
