--- src/main/java/chess/dao/connect/DbConnector.java
@@ -1,8 +1,0 @@
-package chess.dao.connect;
-
-import java.sql.Connection;
-
-public interface DbConnector {
-
-    Connection getConnection();
-}
