--- src/main/java/chess/dao/connect/PsConsumer.java
@@ -1,10 +1,0 @@
-package chess.dao.connect;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface PsConsumer {
-
-    void accept(final PreparedStatement preparedStatement) throws SQLException;
-}
