--- src/main/java/chess/dao/connect/RsFunction.java
@@ -1,10 +1,0 @@
-package chess.dao.connect;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface RsFunction<T> {
-
-    T apply(final ResultSet resultSet) throws SQLException;
-}
