+++ src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:13306/chess
+spring.datasource.username=user
+spring.datasource.password=password
+spring.h2.console.enabled=false
