+++ src/test/java/chess/dao/PlayerDaoTest.java
@@ -1,25 +1,23 @@
-package chess.dao.mysql;
+package chess.dao;
 
 import static chess.domain.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
 
-import chess.dao.connect.CustomJdbcTemplate;
-import chess.dao.connect.TestDbConnector;
 import chess.dao.dto.PlayerDto;
 
+@SpringBootTest
+@Transactional
 class PlayerDaoTest {
 
+    @Autowired
     private PlayerDao playerDao;
-
-    @BeforeEach
-    void setUp() {
-        playerDao = new PlayerDao(new CustomJdbcTemplate(new TestDbConnector()));
-    }
 
     @DisplayName("데이터 저장 및 조회가 가능해야 한다.")
     @Test
@@ -33,7 +31,6 @@ class PlayerDaoTest {
                     assertThat(playerDto.getColorName()).isEqualTo(colorName);
                     assertThat(playerDto.getPieces()).isEqualTo(pieces);
                 });
-        playerDao.remove(id);
     }
 
     @DisplayName("데이터를 수정할 수 있어야 한다.")
@@ -46,6 +43,5 @@ class PlayerDaoTest {
 
         final PlayerDto updatedPlayerDto = playerDao.findById(id);
         assertThat(updatedPlayerDto.getPieces()).isEqualTo(expectedPieces);
-        playerDao.remove(id);
     }
 }
