+++ src/test/resources/application.properties
@@ -1,0 +1,11 @@
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:13307/chess
+spring.datasource.username=user
+spring.datasource.password=password
+spring.h2.console.enabled=false
+
+#spring.datasource.url=jdbc:h2:mem:testdb
+#spring.datasource.driverClassName=org.h2.Driver
+#spring.datasource.username=sa
+#spring.datasource.password=
+#spring.h2.console.enabled=true
